Restructure server and config
