+++ .env.example
@@ -1,18 +1,19 @@
-MAIN_URL=
-TEST_URL=
+# LOCAL ENVIRONMENT VARIABLES
+RPC_SERVER=http://127.0.0.1:7545
 
-PROJECT_ID=
-PROJECT_SECRET=
+USER_ADDRESS=0x000
 
-USER_PUBLIC=
-USER_SECRET=
+CONTRACT_CHAT_ADDRESS=0x000
 
-PORT=
-SEED=
-CHAIN_ID=
+# LOCAL GANACHE ENVIRONMENT VARIABLES
+GANACHE_PORT=7545
+GANACHE_SEED=XXXX
+GANACHE_CHAIN_ID=1337
 
+# PROD ENVIRONMENT VARIABLES
+PROD_RPC_SERVER=wss://mainnet.infura.io/ws/v3/XXXX
+PROD_USER_ADDRESS=0x000
 
-
-CONTRACT_CHAT_ADDRESS=
-RPC_SERVER=
-MY_ADDRESS=
+# TEST ENVIRONMENT VARIABLES
+TEST_RPC_SERVER=wss://ropsten.infura.io/ws/v3/XXXX
+TEST_USER_ADDRESS=0x000
